♻️ Move driver.go to internal/wrap
